[level 0] Title: 배열 만들기 5, Time: 2.87 ms, Memory: 96 MB -BaekjoonHub

+++ 프로그래머스/0/181912. 배열 만들기 5/배열 만들기 5.java
@@ -1,10 +1,10 @@
 import java.util.ArrayList;
 
 class Solution {
-    public ArrayList<Long> solution(String[] intStrs, int k, int s, int l) {
-        ArrayList<Long> list = new ArrayList<>();
+    public ArrayList<Integer> solution(String[] intStrs, int k, int s, int l) {
+        ArrayList<Integer> list = new ArrayList<>();
         for (String intStr : intStrs) {
-            long subStr = Long.parseLong(intStr.substring(s,s+l));
+            int subStr = Integer.parseInt(intStr.substring(s,s+l));
             if (subStr>k)
                 list.add(subStr);
         }
